species: every machine hg-engine numbers has a compatibility bit
The personal record keeps a species' machines in four words, 128 bits, of
which HeartGold used 100: TM01 to TM92 and HM01 to HM08. hg-engine at
d0380a487 numbers 340 machines (sMachineMoves in its src/item.c): those 100
first, then HM07 Dive (ORAS), TM00, TM093 to TM100, Scarlet and Violet's
TM100 to TM229 and TR00 to TR99, and its build sets one bit per machine per
species, the move being in the species' MachineMoves or LevelMoves, a form
without a list reading its base's (build_learnsets.py, write_machine_data;
GetMonMachineMoveCompat reads it). import_species.py dropped every bit past
HM08 -- it said so, 14631 of them across 928 added species -- and HeartGold's
own species never had any, so none of the 240 other machines could have been
taught to anything even once an item mapped to it.

The record grows seven words after the full-width abilities, bits 128 to
351 of hg-engine's numbering; bits 100 to 127 are the unused top of the
fourth word. personal.json carries each species' machines past HM08 as
"machines", their places in that numbering, and the template packs them.
GetTMHMCompatBySpeciesAndForm takes a u16 machine place and reads its word
and bit; places past the 340 are refused. The reference's machine list and
rule are import_species.py's (record, machines_past_hm08); the new
import_machines.py writes the field into every record, HeartGold's
alternate forms 496 to 507 by the number the reference files their
learnsets under. 57904 bits in all. The reference has no learnset for 500,
Trash Cloak Wormadam, so it gets none past HM08 there or here.

Nothing maps an item past HM08 to a machine place yet; that is the next
commit. Engine layer: the data is d0380a487's. konefr's learnsets add five
bits to Sudowoodo and Sunflora, which come in a commit of their own, so
test_species_records' two comparisons with konefr's head fail on those two
until then.

tests/newgold/test_machines.py compares the built archive with the JSON bit
for bit for every species, checks Pikachu has TR08 Thunderbolt, and
compiles the reader natively over a record with one bit set at a time: each
of the 340 places reads its own bit and no other. It fails on the old
reader. Both ROMs build; the suite passes but for the ledger's counter and
the two konefr comparisons above.

+++ include/pokemon.h
@@ -211,9 +211,9 @@ void sub_0207213C(BoxPokemon *boxMon, PlayerProfile *playerProfile, u32 pokeball
 void sub_0207217C(Pokemon *mon, PlayerProfile *a1, u32 pokeball, u32 a3, u32 encounterType, enum HeapID heapID);
 void sub_02072190(BoxPokemon *boxMon, PlayerProfile *a1, u32 pokeball, u32 a3, u32 encounterType, enum HeapID heapID);
 void WildMonSetRandomHeldItem(Pokemon *mon, u32 battleType, u32 isCompoundEyes);
-BOOL GetMonTMHMCompat(Pokemon *mon, u8 tmhm);
-BOOL GetBoxMonTMHMCompat(BoxPokemon *boxMon, u8 tmhm);
-BOOL GetTMHMCompatBySpeciesAndForm(u16 species, u32 form, u8 tmhm);
+BOOL GetMonTMHMCompat(Pokemon *mon, u16 tmhm);
+BOOL GetBoxMonTMHMCompat(BoxPokemon *boxMon, u16 tmhm);
+BOOL GetTMHMCompatBySpeciesAndForm(u16 species, u32 form, u16 tmhm);
 void SetMonPersonality(Pokemon *mon, u32 personality);
 u32 ChangePersonalityToNatureGenderAndAbility(u32 pid, u16 species, u8 nature, u8 gender, u8 ability, BOOL gen_mode);
 void LoadMonPersonal(int species, BASE_STATS *personal);
